Make $default user-only; add heuristics type match
$default is now reserved for user-supplied configurations only. When a new
object is being created, this value is checked and used. Otherwise, the
target type is traversed and defaults are created according to its
structure (traversed as Array, Object, or primitive returns).

Heuristics are now used for conforming a target object to a schema. We
are not yet using heuristics for the "any" array match during
validation. This will probably be controlled by ignoreMissing in the
sense that if the heuristics do not match the value of a fully conformant
instance of the target type, then an error will be reported. Otherwise,
partial matches will be considered as valid during validation.