Rework conformation and validation
This has been done to allow for the new "$typeof" property. This
property allows one to define the schema's type as allowing one of the
items specified in an Array set in $typeof. This is to be used in place
of "$type" when multiple types should be stored in a given field.

If both $type and $typeof are specified, then $typeof will take priortity.