Implement $unmarshal
$unmarshal now expects a function that takes in the input data(whatever
that may be) and returns what it should be transformed into. This is
done at the very beginning of validation as well as conformance, so all
subsequent data checks are against the new type. $validation will likely
be the best avenue for checking the true validity of the new object
type.

Closes #4.