Implement arbitrary object key matching
We can now define a schema as such:

```
{
  "$/.*": [String],
  "$/myKey": [Boolean],
}
```

Which will match objects with keys that match the schema key whose values also match the provided array of types. The array value of this functions in the same was as $typeof, in that you provide an array of types that can be matched.

Closes #10.